TOMEE-2283 - address PR feedback: * I've removed the NOTICE/LICENSE files. * Added missing headers to a few files. * I updated the package name... I had forgot to rename the package name... :) * removed value= when * correct, the deplyment name is not needed here. Removed.
