+++ examples/websocket-tls-basic-auth/README.md
@@ -4,7 +4,11 @@ Websocket example project using SSL and basic authentication.
 
 This was originally created for TomEE 8.0.0. using JEE 8 and Websocket API 1.1.
 
-There is an arquillian test that will perform basic authentication using the username and password configured in 
+The example was created with a server to server typo of connection in mind. 
+For Browser to server connections, you will need to refer to your frontend framework of choice but many server 
+side configurations in here can be reused.
+
+There is an Arquillian test that will perform basic authentication using the username and password configured in 
 the _.../conf/tomcat-users.xml_.
 
 The _.../conf/server.xml_ file used in the Arquillian test can serve as a starting point to a real server configuration.
