+++ examples/websocket-tls-basic-auth/src/main/java/org/superbiz/websockets/WebSocketResource.java
@@ -1,4 +1,4 @@
-package api;
+package org.superbiz.websockets;
 /*
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -22,7 +22,7 @@ import javax.websocket.OnOpen;
 import javax.websocket.Session;
 import javax.websocket.server.ServerEndpoint;
 
-@ServerEndpoint(value = "/socket")
+@ServerEndpoint("/socket")
 public class WebSocketResource {
 
     @OnOpen
